// frontend/apps/editor3d/app/src/pages/Page2d/components/features/ToolbarMain/enum.ts
export enum ToolbarMainButtonNames {
  MENU = "MENU",
  SELECT = "SELECT",
  ADD_TEXT = "ADD_TEXT",
  ADD_IMAGE = "ADD_IMAGE",
  ADD_VIDEO = "ADD_VIDEO",
  ADD_CIRCLE = "ADD_CIRCLE",
  ADD_TRIANGLE = "ADD_TRIANGLE",
  ADD_SQUARE = "ADD_SQUARE",
  PAINT = "PAINT",
  ERASER = "ERASER",
  AI_STYLIZE = "AI_STYLIZE",
  UNDO = "UNDO",
  REDO = "REDO",
  SAVE = "SAVE",
  PREVIEW = "PREVIEW",
  DOWNLOAD = "DOWNLOAD",
  BACKGROUND = "BACKGROUND",

  // 4o generation
  GENERATE = "GENERATE",
}